pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chief-executive-agent"
version = "0.1.0"
description = "A self-improving AI operator that runs locally, does real tasks on your computer, and calls in the big models only when it needs help."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "chromadb>=0.5.0",
    "click>=8.1.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "playwright>=1.44.0",
    "watchdog>=4.0.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
